automatically backed up by learn

+++ index.js
@@ -13,5 +13,5 @@ function happyHolidayTo(holiday, name){
 }
 
 function holidayCountdown(days, holiday){
-  return `It's ${holiday} days until ${days}`
+  return `It's ${holiday} days until ${days}.toInt()`
 }
